+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - 10-bit addressing mode for I2C traits.
+- Added `blocking::dac::SingleChannelDac` trait for digitial-to-analog-converters.
 
 ### Changed